Issue #96. Give `Renderable2dInterface` the AtomicInteger for unique id

+++ dsol-animation-gis-esri/src/main/java/nl/tudelft/simulation/dsol/animation/gis/esri/EsriRenderable2d.java
@@ -224,7 +224,7 @@ public class EsriRenderable2d implements GisRenderable2d
     }
 
     @Override
-    public long getId()
+    public int getId()
     {
         return -1; // drawn before the rest in case all z-values are the same
     }
