+++ dsol-animation-gis-osm/src/main/java/nl/tudelft/simulation/dsol/animation/gis/osm/OsmRenderable2d.java
@@ -223,7 +223,7 @@ public class OsmRenderable2d implements GisRenderable2d
     }
 
     @Override
-    public long getId()
+    public int getId()
     {
         return -1; // drawn before the rest in case all z-values are the same
     }
